+++ components/UI/TextButton.js
@@ -1,6 +1,6 @@
 import styled from 'styled-components';
 
-import ButtonInline from './ButtonInline';
+import ButtonInline from './buttons/ButtonInline';
 
 const TextButton = styled(ButtonInline)`
   /* NOTE: superseded by /components/UI/ButtonInline.js; leaving this here, for now, *
