+++ lib/Matrix.js
@@ -24,6 +24,7 @@ function useMatrixProvider(auth) {
 
     const buildRoomObject = useCallback((roomId) => {
         const room = matrixClient.getRoom(roomId);
+        if (!room) return;
 
         let avatar;
