+++ pages/etherpad/actions/AddExistingPad.js
@@ -3,7 +3,7 @@ import getConfig from 'next/config';
 import { useRouter } from 'next/router';
 import { useTranslation } from 'react-i18next';
 
-import ButtonPrimary from '../../../components/UI/ButtonPrimary';
+import ButtonPrimary from '../../../components/UI/buttons/ButtonPrimary';
 import Form from '../../../components/UI/Form';
 import Input from '../../../components/UI/Input';
 import ErrorMessage from '../../../components/UI/ErrorMessage';
