+++ pages/etherpad/actions/CreateAuthoredPad.js
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
 
-import ButtonPrimary from '../../../components/UI/ButtonPrimary';
+import ButtonPrimary from '../../../components/UI/buttons/ButtonPrimary';
 import Form from '../../../components/UI/Form';
 import Input from '../../../components/UI/Input';
 import LoadingSpinnerInline from '../../../components/UI/LoadingSpinnerInline';
