+++ pages/spacedeck/[[...roomId]].js
@@ -6,7 +6,7 @@ import { useRouter } from 'next/router';
 import { logger } from 'matrix-js-sdk/lib/logger';
 import { DeleteBinIcon } from '@remixicons/react/line';
 
-import ButtonInline from '../../components/UI/ButtonInline';
+import ButtonInline from '../../components/UI/buttons/ButtonInline';
 import LoadingSpinner from '../../components/UI/LoadingSpinner';
 import LoadingSpinnerInline from '../../components/UI/LoadingSpinnerInline';
 import { useAuth } from '../../lib/Auth';
